requirements.txt
requests
flask
peewee
gunicorn
openai
python-dotenv